[fix] : 로그아웃, 탈퇴api 수정-69 - 로그아웃 API 수정 - 탈퇴 API 수정 69 feat
