+++ TellingMe/tellingMe/data/user/api/UserAPI.swift
@@ -11,8 +11,6 @@ import Moya
 enum UserAPITarget {
     case getUserInfo
     case updateUserInfo(UpdateUserInfoRequest)
-    case withdrawalUser
-    case logout
 }
 
 extension UserAPITarget: TargetType {
@@ -31,10 +29,6 @@ extension UserAPITarget: TargetType {
             return "api/user"
         case .updateUserInfo:
             return "api/user/update"
-        case .withdrawalUser:
-            return "api/user/withdraw"
-        case .logout:
-            return "api/user/logout"
         }
     }
 
@@ -42,8 +36,6 @@ extension UserAPITarget: TargetType {
         switch self {
         case .updateUserInfo:
             return .patch
-        case .withdrawalUser, .logout:
-            return .post
         default:
             return .get
         }
@@ -82,30 +74,6 @@ struct UserAPI: Networkable {
     static func updateUserInfo(request: UpdateUserInfoRequest, completion: @escaping(Result<UserInfoResponse?, APIError>) -> Void) {
         do {
             try makeAuthorizedProvider().request(.updateUserInfo(request), dtoType: UserInfoResponse.self, completion: completion)
-        } catch APIError.tokenNotFound {
-            completion(.failure(APIError.tokenNotFound))
-        } catch APIError.errorData(let error) {
-            completion(.failure(APIError.errorData(error)))
-        } catch {
-            completion(.failure(APIError.other(error)))
-        }
-    }
-
-    static func withdrawalUser(completion: @escaping(Result<EmptyResponse?, APIError>) -> Void) {
-        do {
-            try makeAuthorizedProvider().request(.withdrawalUser, dtoType: EmptyResponse.self, completion: completion)
-        } catch APIError.tokenNotFound {
-            completion(.failure(APIError.tokenNotFound))
-        } catch APIError.errorData(let error) {
-            completion(.failure(APIError.errorData(error)))
-        } catch {
-            completion(.failure(APIError.other(error)))
-        }
-    }
-    
-    static func logout(completion: @escaping(Result<EmptyResponse?, APIError>) -> Void) {
-        do {
-            try makeAuthorizedProvider().request(.logout, dtoType: EmptyResponse.self, completion: completion)
         } catch APIError.tokenNotFound {
             completion(.failure(APIError.tokenNotFound))
         } catch APIError.errorData(let error) {
